added support for reading float debug properties from a node

+++ graal/com.oracle.graal.graph/src/com/oracle/graal/graph/NodeClass.java
@@ -729,8 +729,8 @@ public final class NodeClass extends FieldIntrospection {
                     value = unsafe.getLong(node, dataOffsets[i]);
                 } else if (type == Boolean.TYPE) {
                     value = unsafe.getBoolean(node, dataOffsets[i]);
-                } else if (type == Long.TYPE) {
-                    value = unsafe.getLong(node, dataOffsets[i]);
+                } else if (type == Float.TYPE) {
+                    value = unsafe.getFloat(node, dataOffsets[i]);
                 } else if (type == Double.TYPE) {
                     value = unsafe.getDouble(node, dataOffsets[i]);
                 } else {
